+++ src/imageui.h
@@ -51,9 +51,6 @@ double imageui_get_scale(Imageui *imageui);
 void imageui_get_mouse_position(Imageui *imageui,
 	double *image_x, double *image_y);
 
-gboolean imageui_snap_point(Imageui *imageui, int x, int y, int *sx, int *sy);
-gboolean imageui_snap_rect(Imageui *imageui, VipsRect *in, VipsRect *out);
-
 double imageui_get_zoom(Imageui *imageui);
 void imageui_bestfit(Imageui *imageui);
 void imageui_magin(Imageui *imageui);