+++ CHANGELOG.md
@@ -5,7 +5,7 @@ master
 - much better performance for images which are very tall and thin or very wide 
   and narrow
 - better antialiasing [Linden6]
-- copy over improvements from nip4 development
+- copy over improvements from nip4 development, including OME handling
 
 ## 3.1.0 28/10/24